Migrate modules, networks, components, data, vector, rollouts to strict ty
Wave 1 of the full-strict migration: 483 diagnostics fixed for real
across the network layer and env plumbing; these directories are off
the legacy suppression list and now block CI.

Real bugs surfaced by the checker, fixed here:
- rollouts: collect_rollouts/collect_rollouts_recurrent were publicly
  annotated -> list[float] but return a 5-tuple; AsyncPettingZooVecEnv
  step_async declared per-agent lists where workers consume dicts.
- modules: change_kernel_size could write nested ((k,k),(k,k)) tuples
  into the Conv spec and crash a rebuild; multi_input
  init_weights_gaussian had a latent TypeError path for LSTM encoders;
  gpt loss targets took a CPU round-trip via .type(torch.LongTensor).
- gpt KV-cache annotated tuple[Tensor] but unpacked as two tensors.
- networks: StochasticActor.forward declared a 2-tuple, returns 3;
  RolloutBuffer optionals declared non-optional; language_environment
  used list[str, float | None].
- ppo: the recurrent hidden-state family (get_initial_hidden_state,
  get_action, _get_action_and_values, evaluate_actions) untangled from
  dict[str, ArrayOrTensor] to the dict[str, Tensor] it actually is,
  and _get_action_and_values takes the preprocessed TorchObsType.

Where the correct types are blocked by third-party stubs (gymnasium's
CloudpickleWrapper, tensordict's broad __getitem__ union) or genuine
metaprogramming, the sites carry documented casts/ignores.

+++ agilerl/modules/custom_components.py
@@ -230,7 +230,7 @@ class SimbaResidualBlock(nn.Module):
     :param hidden_size: Hidden size of the residual block
     :type hidden_size: int
     :param scale_factor: Scale factor, defaults to 4
-    :type scale_factor: float
+    :type scale_factor: int
     :param device: Device, defaults to "cpu"
     :type device
     """
@@ -238,7 +238,7 @@ class SimbaResidualBlock(nn.Module):
     def __init__(
         self,
         hidden_size: int,
-        scale_factor: float = 4,
+        scale_factor: int = 4,
         device: DeviceType = "cpu",
     ) -> None:
         super().__init__()
